+++ src/main/java/com/example/book_your_seat/queue/util/QueueJwtUtil.java
@@ -1,31 +1,20 @@
 package com.example.book_your_seat.queue.util;
 
-import io.jsonwebtoken.ExpiredJwtException;
+import com.example.book_your_seat.common.util.JwtUtil;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import io.jsonwebtoken.security.SignatureException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.util.Date;
 
-import static com.example.book_your_seat.common.util.JwtConst.*;
-
 @Slf4j
 @Component
-public class QueueJwtUtil {
-    private final SecretKey secretKey;
+public class QueueJwtUtil extends JwtUtil {
     private final Integer expirationTime;
-    private static final String SIGNATURE_ALGORITHM = Jwts.SIG.HS256.key().build().getAlgorithm();
-
     QueueJwtUtil(@Value("${jwt.secret}") String secretKey, @Value("${jwt.queue_expiration_time}") Integer expirationTime) {
-        this.secretKey = new SecretKeySpec(secretKey.getBytes(StandardCharsets.UTF_8), SIGNATURE_ALGORITHM);
+        super(secretKey);
         this.expirationTime = expirationTime;
     }
 
@@ -39,21 +28,6 @@ public class QueueJwtUtil {
                 .expiration(Date.from(expiredAt))
                 .signWith(secretKey)
                 .compact();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token);
-            return true;
-        } catch (MalformedJwtException e) {
-            throw new IllegalStateException(INVALID_JWT);
-        } catch (ExpiredJwtException e) {
-            throw new IllegalStateException(EXPIRED_JWT);
-        } catch (UnsupportedJwtException | SignatureException e) {
-            throw new IllegalStateException(UNSUPPORTED_JWT);
-        } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException(EMPTY_JWT);
-        }
     }
 
     public Long getUserIdByToken(String token) {
